requirements.txt
pdf2image
pillow
questionary
pypdf
click
opencv-python
numpy
pytesseract
ocrmypdf
pikepdf